extract rules for template formatting
